move return_statements module to the correct module
