+++ core_lang/src/parse_tree/mod.rs
@@ -3,6 +3,7 @@ mod code_block;
 mod declaration;
 mod expression;
 mod literal;
+mod return_statement;
 mod use_statement;
 mod while_loop;
 
@@ -11,5 +12,6 @@ pub(crate) use code_block::*;
 pub(crate) use declaration::*;
 pub(crate) use expression::*;
 pub(crate) use literal::Literal;
+pub(crate) use return_statement::*;
 pub(crate) use use_statement::{ImportType, UseStatement};
 pub(crate) use while_loop::WhileLoop;
